local-chatbot/requirements.txt
fastapi
uvicorn[standard]
ollama
cryptography
pywebview
